Small fixes for -Wall -Wextra -Werror

+++ drivers/screen.c
@@ -100,7 +100,7 @@ void disable_cursor(){
 void scroll_screen(size_t rows){
 	memmove((void *) VGA_MEM, (void *) (rows*VGA_WIDTH*2 + VGA_MEM), (VGA_WIDTH*VGA_HEIGHT - rows*VGA_WIDTH)*2);
 	uint16_t *ptr = (uint16_t *) VGA_MEM + (VGA_WIDTH*VGA_HEIGHT - rows*VGA_WIDTH);
-	for (int i = 0; i < rows * VGA_WIDTH; ++i){
+	for (int i = 0; i < (int) rows * VGA_WIDTH; ++i){
 		ptr[i] = (terminal_color << 8) | 0;
 	}
 }
